getting NFD alm example instead of first one
until now we took the first alm example from the list and applied it as a NFD cr.
now because the elements in the array are not in the same order, it caused the NFD cr not to be applied.
This commits finding the appropriate NFD object in the list.